Fixed sprites but screen only a fixed size
Fixed sprites but screen only a fixed size

+++ core/src/com/mygdx/game/entities/Accommodation.java
@@ -9,7 +9,7 @@ import com.mygdx.game.serviceProviders.AccommodationEffectProvider;
 import java.util.HashMap;
 
 public class Accommodation extends POI{
-    public Accommodation(int x, int y, int height, int width) {
+    public Accommodation(double x, double y, double height, double width) {
         super(x,y,height,width);
         this.addComponent(new ComponentCollisionEffect(new AccommodationEffectProvider()));
         this.addComponent(new ComponentSprite(new Texture(Gdx.files.internal("accommodation.png"))));
