+++ core/src/com/mygdx/game/entities/DuckPond.java
@@ -10,7 +10,7 @@ import com.mygdx.game.serviceProviders.LibraryEffectProvider;
 import java.util.HashMap;
 
 public class DuckPond extends POI{
-    public DuckPond(int x, int y, int height, int width) {
+    public DuckPond(double x, double y, double height, double width) {
         super(x,y,height,width);
         this.addComponent(new ComponentCollisionEffect(new DuckPondEffectProvider()));
         this.addComponent(new ComponentSprite(new Texture(Gdx.files.internal("duck_pond.png"))));
